requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
# anthropic  # for Claude version (comment out unused one)